temp/capstone/requirements.txt
PyYAML>=6.0
pytest>=7.0
pytest-asyncio>=0.23
